Updated chart plugins to show more useful info

+++ src/Strem.Plugins.Analytics/Models/Filtering/KeyedMetric.cs
@@ -1,3 +1,3 @@
 ﻿namespace Strem.Plugins.Analytics.Models.Filtering;
 
-public record KeyedMetric(string Key, int Value);
+public record KeyedMetric(string Key, decimal Value);
